Do not load the world if it is not an online session
